Update StudentController & StudentService
Update StudentController to use provide a StudentDTO JSON object rather
than a StudentEntity JSON object.

To make this possible a ModelMapper was added into StudentService to map
between the different classes. This is necessary because the database
requires an instance of the entity class rather than the DTO class.

Of course the tests of StudentController were also updated to reflect
this change.